+++ kafka-streams/kafka-streams/src/main/scala/com/twitter/finatra/streams/transformer/domain/Watermark.scala
@@ -2,6 +2,10 @@ package com.twitter.finatra.streams.transformer.domain
 
 import com.twitter.finatra.streams.converters.time._
 
+object Watermark {
+  val unknown: Watermark = Watermark(0L)
+}
+
 case class Watermark(timeMillis: Long) extends AnyVal {
 
   override def toString: String = {