+++ kafka-streams/kafka-streams/src/main/scala/com/twitter/finatra/streams/transformer/watermarks/internal/WatermarkManager.scala
@@ -4,14 +4,31 @@ import com.twitter.finatra.streams.transformer.OnWatermark
 import com.twitter.finatra.streams.transformer.domain.{Time, Watermark}
 import com.twitter.finatra.streams.transformer.watermarks.WatermarkAssignor
 import com.twitter.inject.Logging
+import org.apache.kafka.streams.processor.TaskId
 
+/**
+ * WatermarkManager coordinates with a Transformers WatermarkAssignor to keep track of the latest assigned
+ * watermark and the last emitted watermark.
+ *
+ * @param taskId TaskId of the FinatraTransformer being managed used for internal logging
+ * @param transformerName Transformer name of the FinatraTransformer being managed used for internal logging
+ * @param onWatermark OnWatermark callback which is called when a new watermark is emitted
+ * @param watermarkAssignor The WatermarkAssignor used in the FinatraTransformer being managed.
+ * @param emitWatermarkPerMessage Whether to check if a new watermark needs to be emitted after each
+ *                                message is read in onMessage. If false, callOnWatermarkIfChanged must
+ *                                be called to check if a new watermark is to be emitted.
+ * @tparam K Message key for the FinatraTransformer being managed
+ * @tparam V Message value for the FinatraTransformer being managed
+ */
 class WatermarkManager[K, V](
+  taskId: TaskId,
+  transformerName: String,
   onWatermark: OnWatermark,
   watermarkAssignor: WatermarkAssignor[K, V],
   emitWatermarkPerMessage: Boolean)
     extends Logging {
 
-  @volatile private var lastEmittedWatermark = Watermark(0L)
+  @volatile private var lastEmittedWatermark = Watermark.unknown
 
   /* Public */
 
@@ -26,16 +43,17 @@ class WatermarkManager[K, V](
   def onMessage(messageTime: Time, topic: String, key: K, value: V): Unit = {
     watermarkAssignor.onMessage(topic = topic, timestamp = messageTime, key = key, value = value)
 
-    if (emitWatermarkPerMessage) {
+    if (lastEmittedWatermark == Watermark.unknown || emitWatermarkPerMessage) {
       callOnWatermarkIfChanged()
     }
   }
 
   def callOnWatermarkIfChanged(): Unit = {
-    val currentWatermark = watermarkAssignor.getWatermark
-    if (currentWatermark.timeMillis > lastEmittedWatermark.timeMillis) {
-      onWatermark.onWatermark(currentWatermark)
-      setLastEmittedWatermark(currentWatermark)
+    val latestAssignedWatermark = watermarkAssignor.getWatermark
+    trace(s"callOnWatermarkIfChanged $transformerName $taskId $latestAssignedWatermark")
+    if (latestAssignedWatermark.timeMillis > lastEmittedWatermark.timeMillis) {
+      onWatermark.onWatermark(latestAssignedWatermark)
+      setLastEmittedWatermark(latestAssignedWatermark)
     }
   }